Removed various if (1) and if (0) statements and changed them in #if 0 and other better suited code that does not generate warnings.

+++ include/extern.h
@@ -1590,7 +1590,7 @@ E void FDECL(change_mon_ability_scores, (struct monst*, int, int));
 E struct obj* FDECL(mongets, (struct monst *, int));
 E int FDECL(mongets_return_enchantment, (struct monst*, int));
 E struct obj* FDECL(mon_gets_noinit_item, (struct monst*, int,  int));
-E int FDECL(mongetsgold, (struct monst*, long));
+E long FDECL(mongetsgold, (struct monst*, long));
 E int FDECL(golemhp, (int));
 E boolean FDECL(peace_minded, (struct permonst *));
 E void FDECL(set_malign, (struct monst *));
